+++ packages/apollo-server-plugin-operation-registry/src/agent.ts
@@ -12,7 +12,6 @@ import loglevel from 'loglevel';
 import { Response } from 'node-fetch';
 import { InMemoryLRUCache } from 'apollo-server-caching';
 import { fetchIfNoneMatch } from './fetchIfNoneMatch';
-import { ValueOrPromise } from "apollo-server-plugin-base";
 import { OperationManifest } from "./ApolloServerPluginOperationRegistry";
 
 const DEFAULT_POLL_SECONDS: number = 30;
@@ -25,10 +24,6 @@ export interface AgentOptions {
   engine: any;
   store: InMemoryLRUCache;
   schemaTag: string;
-  willUpdateManifest?: (
-    newManifest?: OperationManifest,
-    oldManifest?: OperationManifest,
-  ) => ValueOrPromise<OperationManifest>;
 }
 
 type SignatureStore = Set<string>;
@@ -47,7 +42,6 @@ export default class Agent {
   public _timesChecked: number = 0;
 
   private lastOperationSignatures: SignatureStore = new Set();
-  private lastManifest?: OperationManifest;
   private readonly options: AgentOptions = Object.create(null);
 
   constructor(options: AgentOptions) {
@@ -98,14 +92,8 @@ export default class Agent {
     try {
       await pulse();
     } catch (err) {
-      // Update the manifest to trigger `willUpdateManifest(newManifest: undefined, oldManifest: undefined)`
-      await this.updateManifest();
       console.error(
-        `The operation manifest could not be fetched. Retries will continue.${
-          this.lastManifest
-            ? ''
-            : ' Requests will be forbidden until the manifest is fetched.'
-        }`,
+        'The operation manifest could not be fetched. Retries will continue, but requests will be forbidden until the manifest is fetched.',
         err.message || err,
       );
     }
@@ -308,37 +296,19 @@ export default class Agent {
       });
   }
 
-  public async updateManifest(manifest?: OperationManifest) {
+  public async updateManifest(manifest: OperationManifest) {
     if (
-      manifest &&
-      (manifest.version !== 2 || !Array.isArray(manifest.operations))
+      !manifest ||
+      manifest.version !== 2 ||
+      !Array.isArray(manifest.operations)
     ) {
       throw new Error('Invalid manifest format.');
     }
 
-    const returnedManifest =
-      (await (this.options.willUpdateManifest &&
-        this.options.willUpdateManifest(manifest, this.lastManifest))) ||
-      manifest;
-
-    if (returnedManifest && !Array.isArray(returnedManifest.operations)) {
-      throw new Error(
-        "Invalid manifest format. Manifest's operations must be an array",
-      );
-    }
-
-    if (returnedManifest) {
-      this.updateOperationStore(returnedManifest.operations);
-    }
-
-    this.lastManifest = manifest;
-  }
-
-  private updateOperationStore(operations: OperationManifest['operations']) {
     const incomingOperations: Map<string, string> = new Map();
     const replacementSignatures: SignatureStore = new Set();
 
-    for (const { signature, document } of operations) {
+    for (const { signature, document } of manifest.operations) {
       incomingOperations.set(signature, document);
       // Keep track of each operation in this manifest so we can store it
       // for comparison after the next fetch.